fixing ReadRequestHeader
